Write initial Tarragon Server Pages grammar
After much refreshing up on and plodding around with LPEG, I have an initial grammar successfully working for TSP, as confirmed by REPL.

It identifies matches for now, but not much more. I also need to refine a few of the lexical rules.

Next up is to split up the fennel-code pattern into three segments (for when captures occur - open, code to be parsed, and close), and then work on updating the patterns into captures that return the structure of the server page as a table to be handled accordingly.