Implement segmenting of server page into distinct HTML and Fennel segments
Went forward and refactored the Lpeg patterns so that when a Tarragon server page is loaded, it returns a sequential table of tagged HTML and Fennel blocks.

The test was updated, it passes so far, successfully segmenting a sample TSP.

Next up is to build convenience functions around the returned table to create a reusable function / closure that allows the page to be served over and over again without have to be re-parsed (memoization). Also, the fennel-code structure needs to be refactored so as to only return the Fennel code, and not the delimiting symbols, though I haven't decided whether I want to do that in the grammar itself or in some separate, later pass over the code again. The returned table would also benefit from retaining line numbers and absolute position, for debugging purposes when the parsed page inevitably contains bugs, though I don't know how soon I want to implement that (I want to return to the other project this repo supports, asap, and am thus antsy).