Update parser to return Fennel code without delimiters; start adding files for other needed components
A small update with an lpeg capture allowed me to easily get rid of those delimiters in the returned table.

I also started creating files for other modules that will be needed, though none have been implemented to any usable extent yet. The parsing code was moved from init to parser.

Next up is to continue on with convenience functions for the parser, and to move forward with an interpreter, and then a memoizer/cacher and an actual rendering test.