Implement proof-of-concept with first server page evaluation
Holy stromboli - what a night. I just spent many hours reading over the Lua docs on chunks and _ENV and _G, and reading up on Fennel's eval function and how it relates and differs. I then spent quite a while thinking about how I was actually going to implement this, and if I can, what philosophies and paradigms I should encourage.

My original intent - since Redbean forks processes that can't influence each other since they clone the end state of the /ZIP/.init.lua - was to have a processed page share the same environment with the function that calls it.

However, this was very misguided, as the loaded chunks are not inlined code, but rather, they are compiled into standalone functions. Global references can be shared, but it is incredibly onerous to share environment, as in both Lua and Fennel you would have to manually provide the references to each chunk that you want to use from the containing environment. NOPE.

Okay, so I realized that the pages are each going to have to be 'distinct functions', where each page is converted to a chunk. Ideally, every bit of code on the page would be considered part of one big 'page function' for that specific page.

However, this is not as trivial as it sounds. Each segment of Fennel code spread throughout a single page can be concatenated into one large string and then fennel.eval'ed. But this doesn't cut it - if the idea is to have an output function that renders text to the page, then you need to be able to intercept the return values of that output function, which isn't possible if you parse it as one big chunk. If you separate them into different chunks though, then they can no longer share variables and other definitions.

I did eventually think up a solution that allows both - I will parse each segment one-by-one, injecting intermediary code, a header, and a footer. The header will establish an output factory that returns a closure. This closure will be the output function. Then each segment will invoke it, and in between each segment, I will inject code that updates the closure or overwrites it with a new one. The footer will then process return values. And the page will be considered one large function - it will accept only one big vararg, so that no specific definition ever needs to be written - users can be certain that any values at all will come in the form of a var arg, and if they want something akin to a function signature, then they can destructure the vararg at the top of the Tarragon server page.

Now, this isn't impossible or even really that difficult at all, but it is a bit onerous, especially since I'm getting quite sleepy. To appease my need for instant gratification, I went ahead and implemented a function in the interpreter that picks up the Fennel code from a page and evaluates it all at once. I know this is exactly contrary to what I just decided to implement, but I just wanted to see something functioning at all before dozing off - I will implement my actual vision in the coming days.

Next up is to begin writing the micro-functions needed to scaffold to the vision I just elaborated, where injecting code will be needed.